+++ frontend/code/components/link.ts
@@ -1,11 +1,13 @@
 import { ComponentId } from '../dataModels';
 import { ComponentBase, ComponentState } from './componentBase';
 import { hijackLinkElement } from '../utils';
+import { applyIcon } from '../designApplication';
 
 export type LinkState = ComponentState & {
     _type_: 'Link-builtin';
     child_text?: string | null;
     child_component?: ComponentId | null;
+    icon?: string | null;
     open_in_new_tab?: boolean;
     targetUrl: string;
 };
@@ -53,16 +55,32 @@ export class LinkComponent extends ComponentBase {
 
         // Child Text?
         if (
-            deltaState.child_text !== undefined &&
-            deltaState.child_text !== null
+            (deltaState.child_text !== undefined &&
+                deltaState.child_text !== null) ||
+            (this.state.child_text !== null && deltaState.icon !== undefined)
         ) {
             // Clear any existing children
             this.removeHtmlChild(latentComponents);
 
+            // Add the icon, if any
+            console.debug(deltaState.icon, this.state.icon);
+            let icon = deltaState.icon ?? this.state.icon;
+
+            if (icon !== null) {
+                let iconElement = document.createElement('div');
+                iconElement.classList.add('rio-text-link-icon');
+                element.appendChild(iconElement);
+
+                applyIcon(iconElement, icon, 'currentColor');
+            }
+
             // Add the new text
+            let child_text = deltaState.child_text ?? this.state.child_text;
+
             let textElement = document.createElement('div');
+            textElement.classList.add('rio-text-link-text');
             element.appendChild(textElement);
-            textElement.textContent = deltaState.child_text;
+            textElement.textContent = child_text;
 
             // Update the CSS classes
             element.classList.add('rio-text-link');